+++ docs/tr1/CHANGELOG.md
@@ -6,6 +6,7 @@
 - fixed extremely large item quantities crashing the game (#2497, regression from 0.3)
 - fixed Lara's meshes not resetting after using the fly cheat (#2565, #2572, regressions from 4.8)
 - fixed guns appearing in Lara's hands if the draw input is held when unarmed and while picking up a gun item (#2577, regressions from 0.8/4.3)
+- fixed being able to play with Lara invisible after using the explosion cheat then the fly cheat (#2584, regression from 4.8)
 
 ## [4.8.3](https://github.com/LostArtefacts/TRX/compare/tr1-4.8.2...tr1-4.8.3) - 2025-02-17
 - fixed some of Lara's speech in the gym not playing in response to player action (#2514, regression from 4.8)